requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
diff-match-patch>=20230430
